Crate osmia_npm

Functions

ctx_json_dump

Get the Osmia context, based on a given initial context.

ctx_json_dump_variable

Get a variable from the Osmia context, based on a given initial context.

ctx_yaml_dump

Get the Osmia context, based on a given initial context.

ctx_yaml_dump_variable

Get a variable from the Osmia context, based on a given initial context.

run

Run Osmia code with the default context

run_ctx

Run Osmia code with a custom context

run_json_ctx

Run Osmia code with a custom JSON context

run_yaml_ctx

Run Osmia code with a custom YAML context